housekeeping and cross browser compatibility stuff
A lot of shuffling of the CSS. Changed several elements in CSS so they
will not highlight when clicked.

Button area now working correctly. A deadButton class is now in in HTML
and styled to work correctly in the CSS file.

Worked on event handlers in JS try and fix some cross browser bugs.