fix (mostly)  firefox problem when context menu opens
When the context menu was opened and the mouse moved over it in firefox
the mouseout event was firing causing the "paste message here..."
message to pop into the box.

These fixes mostly fix that although the mouseout "paste message
here..." functionality may be turned off if an item is clicked on in the
context menu until the user clicks on something in the document. Not
sure how to fix this. Would be nice if there were a contexmenuIsClosed
event but does not look like that exist.