Changes that allow enigmaCore to be used as a Regular Enigma Machine
These changes allow enigmaCore to be used to build a regular Enigma
machine that can talk to the other Enigma machines listed in the about
section.

These changes were carefully added as to not effect enigmaX's current
functionality. enigmaX should still run just as before.

A double stepping feature was added, but is off by default. Double
stepping is needed for the regular Enigma to function properly.

Ring settings were added to the rotors and enigmaCore can change the
ring settings now. By default the rings are left alone and do nothing
but ring settings can now be added to any enigmaCore key.

enigmaX does not use ring settings but it's security could be increased
by using this feature, however the size of it's key would have to be
increased.