Fixes Chrome bug where the X loads weird
The X in the title was loading strange in Chrome. The first time you
loaded it it would be in the middle of the key box. Reloading would make
it correct.

By changing the style of the element to absolute after the page loads
(done in JavaScript instead of the CSS) the problem is fixed.