Mitigates IE textarea font fallback bug
In this bug in IE 10 (and maybe other versions) fonts in textarea do not
fallback correctly to their font-family fonts. A textarea will only take
the first font given it and use it's own font for any other glyphs. This
is very problematic for my Think Ding font, since it needs to be used
along with a regular font.

I asked about this problem on stackoverflow:
http://stackoverflow.com/questions/19442749/textarea-fallback-in-ie

The solution offered was to  use a div and the contenteditable
attribute. This solution seems workable but would introduce so many new
problems it does not seem worth it at this time.

Instead I have tried to mitigate the bug here. The nice thing about this
solution is it should "succeed gracefully" if the bug is ever fixed. It
also doesn't mess with any of the correct functioning in Firefox and
Chrome.

The downside is in IE, in the message box, Think Ding strings will now
be a bit ugly.

(note: there are also a few esthetic fixes in this commit, such as font
size increase in the message box and some corrections that make the
about section look cleaner)