Allows Think Ding to be re-encrypted
These functions allow you to encrypt anything in Think Ding again by
simply putting bracket [] around the Think Ding.

It finds any groups of bracketed Think Ding and processes them into
ASCII and places them in groups back in the message. This allows the
regular Enigma X encryption algorithm to see the message as ASCII and
encrypt it normally.

The process is revered after decryption.

By doing it this way the encryption algorithm is not touched at all.
Instead these functions just interfaces with it. This is not the most
efficient way to do it but it allows the added the functionality without
breaking anything that's already working.

This could use some more testing to make sure it's working all the time.