Adds key encryption - Enigma X is Done!
This adds the final feature to Enigma X. This is version 1.0!

The main addition here is a processTDKeys() which finds any Think Ding
formatted as a key in the message and processes into ASCII to be
encrypted with the message.

The crypt() function in EnigmaXMachine() now strips encrypted Think Ding
messages of any spaces and new lines to make the machine a bit more
forgiving with inputs.

There are a few other additions that make everything work together.