Scrolling in message box and about area
This adds some user friendliness by bringing back some of the scrolling
mechanisms while still maintaining the basic aesthetic of the program.

Now the user can scroll with the mouse wheel in long messages in the
message box, although the scroll bar is still hidden.

In the about section if necessary the scroll bar will appear with normal
web page functionality.

This functionality is mostly accomplished by pushing the scroll bar out
of the div. It's a bit of a hacky way to do it and is likely to cause
bugs in some browsers, but so far works in all IE 10, Firefox.

There was a mall bug in Chrome but chrome has it's own scrollbar
manipulation so that's used to keep the scroll bar hidden.

Safari has not yet been tested. Cross fingers.