Fix blinkRed() and dropping X scroll bar bug
Since the outer message box was made the main border box to create the
scrolling functionality, the blinkRed() function needed to be send the
correct div.

Otherwise this is just a lot of time spent trying to get the X to play
well with the scroll bar in the about Div.