Fix Firefox in Linux context menu bug
When context menu is opened the UI the informed via contextMenuOn not to
fire the default message functionality.  This keeps the default message
from popping up when the user is trying to paste stuff in the box. When
you click on the document again the UI is informed the context menu is
gone which lets the default message functionality return.

The problem is in Firefox on Linux the context menu event fires first
and then the document.onclick is fired for the right click. This causes
the contextMenuOn flag to be turned on then off immediately, messesing
up it's functionality.

I've fixed this by keeping it from turning off if it was the right mouse
button that was clicked.

This problem is also happening in Firefox on OSX. I haven't tested it
yet but I think this commit will fix that also.

I'm also going to start appending the readme file so that the version
number listed there has the commit number appended on it. This will be a
simple way to keep track of versions in the wild.

However in