Adds extra line ending to ThinkDing stripping
This helps make sure the ThinkDing is stripped if it is put in the box.

This is to fix a bug that was happening when a ThinkDing message was
pasted in from Google+.

Google+ was adding a bite order mark (uFEFF) at the end of the line and
that was breaking the stripping function and not allowing Enigma X to
decrypt the message.  This fixes that, although it points the way to
other problems that can occur.

Probably even more possible invisible chars should be added to this
regular expression.