Adds fade in on load of page
This is much smoother than any of the other solutions and looks cool. Yet
to be tested on the slow server so we will see how it fairs.

Also small change of "Generating Key" to "Generating Random Key" message
that is displayed in IE. This is in preperation for an idea to implement the
little key generation animation for all browsers.